fix(dashboard): Create Combobox componet to replace AutoComplete(v2)
