--- dashboard/src/components/form/FormAutocomplete/index.ts
@@ -1,7 +1,0 @@
-export { default as FormAutocomplete } from './FormAutocomplete';
-export {
-  default as FormAutocompleteCore,
-  type FormAutocompleteChangeMeta,
-  type FormAutocompleteChangeSource,
-  type FormAutocompleteOption,
-} from './FormAutocompleteCore';
