+++ dashboard/src/components/form/FormCombobox/FormCombobox.tsx
@@ -1,39 +1,19 @@
-import { Check, ChevronsUpDown } from 'lucide-react';
-import type { ForwardedRef, ReactNode } from 'react';
-import { forwardRef, useState } from 'react';
+import type { ReactNode } from 'react';
 import type { Control, FieldPath, FieldValues } from 'react-hook-form';
-import { mergeRefs } from 'react-merge-refs';
-import { Button } from '@/components/ui/v3/button';
+import Combobox, { type ComboboxOption } from '@/components/ui/v3/combobox';
 import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from '@/components/ui/v3/command';
-import {
-  FormControl,
   FormDescription,
   FormField,
   FormItem,
   FormLabel,
   FormMessage,
 } from '@/components/ui/v3/form';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/v3/popover';
-import { cn, isNotEmptyValue } from '@/lib/utils';
+import { cn } from '@/lib/utils';
+
+export type { ComboboxOption as FormComboboxOption };
 
 const comboboxTriggerClasses =
   'aria-[invalid=true]:border-red-500 aria-[invalid=true]:focus:border-red-500 aria-[invalid=true]:focus:ring-red-500';
-
-export interface FormComboboxOption {
-  value: string;
-  label: ReactNode;
-}
 
 interface FormComboboxProps<
   TFieldValues extends FieldValues = FieldValues,
@@ -49,147 +29,80 @@ interface FormComboboxProps<
   helperText?: ReactNode | null;
   disabled?: boolean;
   searchPlaceholder?: string;
-  emptyText?: string;
-  options: FormComboboxOption[];
+  emptyText?: ReactNode;
+  options: ComboboxOption[];
+  filter?: (value: string, search: string, keywords?: string[]) => number;
   'data-testid'?: string;
 }
 
-function FormComboboxImpl<
+export default function FormCombobox<
   TFieldValues extends FieldValues = FieldValues,
   TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
->(
-  {
-    control,
-    name,
-    label,
-    placeholder,
-    className,
-    containerClassName,
-    inline,
-    helperText,
-    disabled,
-    options,
-    searchPlaceholder = 'Search...',
-    emptyText = 'No results found.',
-    'data-testid': dataTestId,
-  }: FormComboboxProps<TFieldValues, TName>,
-  ref?: ForwardedRef<HTMLButtonElement>,
-) {
-  // biome-ignore lint/correctness/useHookAtTopLevel: forwardRef render function with generic type cast
-  const [open, setOpen] = useState(false);
-
+>({
+  control,
+  name,
+  label,
+  placeholder,
+  className,
+  containerClassName,
+  inline,
+  helperText,
+  disabled,
+  searchPlaceholder,
+  emptyText,
+  options,
+  filter,
+  'data-testid': dataTestId,
+}: FormComboboxProps<TFieldValues, TName>) {
   return (
     <FormField
       control={control}
       name={name}
-      render={({ field }) => {
-        const { ref: fieldRef, onChange, value } = field;
-
-        const selectedOption = options.find((opt) => opt.value === value);
-        const selectedLabel = selectedOption?.label ?? value;
-
-        return (
-          <FormItem
-            className={cn(
-              { 'flex w-full items-center gap-4 py-3': inline },
-              containerClassName,
-            )}
+      render={({ field, fieldState }) => (
+        <FormItem
+          className={cn(
+            { 'flex w-full items-center gap-4 py-3': inline },
+            containerClassName,
+          )}
+        >
+          <FormLabel
+            className={cn({
+              'w-52 max-w-52 flex-shrink-0': inline,
+              'mt-2 self-start': inline && !!helperText,
+            })}
           >
-            <FormLabel
-              className={cn({
-                'w-52 max-w-52 flex-shrink-0': inline,
-                'mt-2 self-start': inline && !!helperText,
-              })}
-            >
-              {label}
-            </FormLabel>
-            <div
-              className={cn({
-                'flex w-[calc(100%-13.5rem)] max-w-[calc(100%-13.5rem)] flex-col gap-2':
-                  inline,
-              })}
-            >
-              <Popover open={open} onOpenChange={setOpen}>
-                <PopoverTrigger asChild>
-                  <FormControl>
-                    <Button
-                      type="button"
-                      variant="outline"
-                      role="combobox"
-                      className={cn(
-                        comboboxTriggerClasses,
-                        'w-full justify-between',
-                        !selectedLabel &&
-                          !isNotEmptyValue(value) &&
-                          'text-muted-foreground',
-                        className,
-                      )}
-                      disabled={disabled}
-                      ref={mergeRefs([fieldRef, ref])}
-                      data-testid={dataTestId}
-                    >
-                      {selectedLabel ||
-                        (isNotEmptyValue(value) ? value : placeholder)}
-                      <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                    </Button>
-                  </FormControl>
-                </PopoverTrigger>
-                <PopoverContent className="max-h-[var(--radix-popover-content-available-height)] w-[var(--radix-popover-trigger-width)] p-0">
-                  <Command>
-                    <CommandInput
-                      placeholder={searchPlaceholder}
-                      className="h-9"
-                      disabled={disabled}
-                    />
-                    <CommandList>
-                      <CommandEmpty>{emptyText}</CommandEmpty>
-                      <CommandGroup>
-                        {options.map((option) => (
-                          <CommandItem
-                            key={option.value}
-                            value={option.value}
-                            onSelect={() => {
-                              onChange(option.value);
-                              setOpen(false);
-                            }}
-                          >
-                            {option.label}
-                            <Check
-                              className={cn(
-                                'ml-auto size-4',
-                                option.value === value
-                                  ? 'opacity-100'
-                                  : 'opacity-0',
-                              )}
-                            />
-                          </CommandItem>
-                        ))}
-                      </CommandGroup>
-                    </CommandList>
-                  </Command>
-                </PopoverContent>
-              </Popover>
-              {!!helperText && (
-                <FormDescription className="break-all px-[1px]">
-                  {helperText}
-                </FormDescription>
-              )}
-              <FormMessage />
-            </div>
-          </FormItem>
-        );
-      }}
+            {label}
+          </FormLabel>
+          <div
+            className={cn({
+              'flex w-[calc(100%-13.5rem)] max-w-[calc(100%-13.5rem)] flex-col gap-2':
+                inline,
+            })}
+          >
+            <Combobox
+              ref={field.ref}
+              value={field.value ?? null}
+              onChange={field.onChange}
+              onBlur={field.onBlur}
+              options={options}
+              filter={filter}
+              placeholder={placeholder}
+              searchPlaceholder={searchPlaceholder}
+              emptyText={emptyText}
+              disabled={disabled}
+              className={cn(comboboxTriggerClasses, className)}
+              aria-invalid={!!fieldState.error}
+              data-testid={dataTestId}
+            />
+            {!!helperText && (
+              <FormDescription className="break-all px-[1px]">
+                {helperText}
+              </FormDescription>
+            )}
+            <FormMessage />
+          </div>
+        </FormItem>
+      )}
     />
   );
 }
-
-const FormCombobox = forwardRef(FormComboboxImpl) as <
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
->(
-  props: FormComboboxProps<TFieldValues, TName> & {
-    ref?: ForwardedRef<HTMLButtonElement>;
-  },
-) => ReturnType<typeof FormComboboxImpl>;
-
-export default FormCombobox;
