+++ dashboard/src/components/form/FormFreeCombobox/index.ts
@@ -1,0 +1,4 @@
+export {
+  default as FormFreeCombobox,
+  type FormFreeComboboxOption,
+} from './FormFreeCombobox';
